src/test/javascript/spec/app/entities/produkt/produkt.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import { ProduktService } from 'app/entities/produkt/produkt.service';
import { IProdukt, Produkt } from 'app/shared/model/produkt.model';
import { StatusProdukt } from 'app/shared/model/enumerations/status-produkt.model';
import { ProduktKategoriaEnum } from 'app/shared/model/enumerations/produkt-kategoria-enum.model';

describe('Service Tests', () => {
  describe('Produkt Service', () => {
    let injector: TestBed;
    let service: ProduktService;
    let httpMock: HttpTestingController;
    let elemDefault: IProdukt;
    let expectedResult;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = {};
      injector = getTestBed();
      service = injector.get(ProduktService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Produkt(0, 'AAAAAAA', 0, 'AAAAAAA', StatusProdukt.DOSTEPNY, 'AAAAAAA', 0, ProduktKategoriaEnum.GRZEJNIKI);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: elemDefault });
      });

      it('should create a Produkt', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new Produkt(null))
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should update a Produkt', () => {
        const returnedFromService = Object.assign(
          {
            nazwa: 'BBBBBB',
            cena: 1,
            opis: 'BBBBBB',
            status: 'BBBBBB',
            zdjecie: 'BBBBBB',
            stan: 1,
            kategoria: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject({ body: expected });
      });

      it('should return a list of Produkt', () => {
        const returnedFromService = Object.assign(
          {
            nazwa: 'BBBBBB',
            cena: 1,
            opis: 'BBBBBB',
            status: 'BBBBBB',
            zdjecie: 'BBBBBB',
            stan: 1,
            kategoria: 'BBBBBB'
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query(expected)
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Produkt', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
